weavy: remove unused imports and frameworks
This commit removes unused imports such as Foundation and ObjectiveC
from the Weavy files.

It also removes the unused dependencies concerning NSObject-Rx and
RxSwiftExt's pausable. Since these files (or portions of code) are
directly added in the project there is no need to bootstrap them
with Carthage anymore, or to keep broken references.
However, in order to satisfy the licenses, the copyright and license
notices were added to the project.

Since some functions comments have changed, the documentation has also
been regenerated.